# perf(python): Add python benchmark test