--- a/integration_tests/cpython_benchmark/fury_benchmark.py
+++ b/integration_tests/cpython_benchmark/fury_benchmark.py
@@ -1,0 +1,121 @@
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
+
+import datetime
+import random
+import sys
+from pyfury import Fury, Language
+import pyperf
+
+
+# The benchmark case is rewritten from pyperformance bm_pickle
+# https://github.com/python/pyperformance/blob/main/pyperformance/data-files/benchmarks/bm_pickle/run_benchmark.py
+DICT = {
+    "ads_flags": 0,
+    "age": 18,
+    "birthday": datetime.date(1980, 5, 7),
+    "bulletin_count": 0,
+    "comment_count": 0,
+    "country": "BR",
+    "encrypted_id": "G9urXXAJwjE",
+    "favorite_count": 9,
+    "first_name": "",
+    "flags": 412317970704,
+    "friend_count": 0,
+    "gender": "m",
+    "gender_for_display": "Male",
+    "id": 302935349,
+    "is_custom_profile_icon": 0,
+    "last_name": "",
+    "locale_preference": "pt_BR",
+    "member": 0,
+    "tags": ["a", "b", "c", "d", "e", "f", "g"],
+    "profile_foo_id": 827119638,
+    "secure_encrypted_id": "Z_xxx2dYx3t4YAdnmfgyKw",
+    "session_number": 2,
+    "signup_id": "201-19225-223",
+    "status": "A",
+    "theme": 1,
+    "time_created": 1225237014,
+    "time_updated": 1233134493,
+    "unread_message_count": 0,
+    "user_group": "0",
+    "username": "collinwinter",
+    "play_count": 9,
+    "view_count": 7,
+    "zip": "",
+}
+
+TUPLE = (
+    [
+        265867233,
+        265868503,
+        265252341,
+        265243910,
+        265879514,
+        266219766,
+        266021701,
+        265843726,
+        265592821,
+        265246784,
+        265853180,
+        45526486,
+        265463699,
+        265848143,
+        265863062,
+        265392591,
+        265877490,
+        265823665,
+        265828884,
+        265753032,
+    ],
+    60,
+)
+
+
+LIST = [[list(range(10)), list(range(10))] for _ in range(10)]
+
+
+def mutate_dict(orig_dict, random_source):
+    new_dict = dict(orig_dict)
+    for key, value in new_dict.items():
+        rand_val = random_source.random() * sys.maxsize
+        if isinstance(key, (int, bytes, str)):
+            new_dict[key] = type(key)(rand_val)
+    return new_dict
+
+
+random_source = random.Random(5)
+DICT_GROUP = [mutate_dict(DICT, random_source) for _ in range(3)]
+
+
+def fury_python_object(obj):
+    fury = Fury(language=Language.PYTHON, ref_tracking=True)
+    binary = fury.serialize(obj)
+    fury.deserialize(binary)
+
+
+def micro_benchmark():
+    runner = pyperf.Runner()
+    runner.bench_func("fury_dict", fury_python_object, DICT)
+    runner.bench_func("fury_dict_group", fury_python_object, DICT_GROUP)
+    runner.bench_func("fury_tuple", fury_python_object, TUPLE)
+    runner.bench_func("fury_list", fury_python_object, LIST)
+
+
+if __name__ == "__main__":
+    micro_benchmark()